Only link pulp-agent to goferd init script if it doesn't exist already

+++ pulp.spec
@@ -248,7 +248,9 @@ cp src/pulp/client/yumplugin/pulp-profile-update.py %{buildroot}/usr/lib/yum-plu
 # Pulp and CDS init.d
 mkdir -p %{buildroot}/etc/rc.d/init.d
 cp etc/rc.d/init.d/* %{buildroot}/etc/rc.d/init.d/
-ln -s etc/rc.d/init.d/goferd %{buildroot}/etc/rc.d/init.d/pulp-agent
+if [ ! -e %{buildroot}/etc/rc.d/init.d/pulp-agent ] then
+    ln -s etc/rc.d/init.d/goferd %{buildroot}/etc/rc.d/init.d/pulp-agent
+fi
 
 # Remove egg info
 rm -rf %{buildroot}/%{python_sitelib}/%{name}*.egg-info
